+++ Controller/Backend/DashboardController.php
@@ -1,16 +1,13 @@
 <?php
 
-namespace Opifer\CmsBundle\Controller;
+namespace Opifer\CmsBundle\Controller\Backend;
 
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
 
 class DashboardController extends Controller
 {
     /**
-     * @Route("/", name="opifer.cms.dashboard.view")
-     *
      * @return Response
      */
     public function viewAction()
@@ -18,7 +15,7 @@ class DashboardController extends Controller
         $latestContent = $this->getDoctrine()->getRepository('OpiferCmsBundle:Content')
             ->findLastUpdated(8);
 
-        return $this->render('OpiferCmsBundle:Dashboard:dashboard.html.twig', [
+        return $this->render('OpiferCmsBundle:Backend/Dashboard:dashboard.html.twig', [
             'latest_content' => $latestContent
         ]);
     }
