+++ Resources/views/base.html.twig
@@ -27,7 +27,7 @@
                     <div class="container-fluid">
                         <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
                             <ul class="nav navbar-nav">
-                                <li><a href="{{ path('opifer.cms.dashboard.view') }}"><i class="fa fa-rocket"></i></a></li>
+                                <li><a href="{{ path('opifer_cms_dashboard') }}"><i class="fa fa-rocket"></i></a></li>
                                 <li><a href="{{ path('opifer_content_content_index') }}">{{ "nav.content"|trans|capitalize }}</a></li>
                                 <li><a href="{{ path('opifer_media_media_index') }}"><i class="fa fa-photo visible-xs-block"></i> {{ "nav.media"|trans|capitalize }}</a></li>
                                 <li><a href="{{ path('opifer_content_directory_index') }}"><i class="fa fa-folder-o visible-xs-block"></i> {{ "nav.directory"|trans|capitalize }}</a></li>
